README.md
This is a small app idea that I've used for learning the MERN stack. It'll utalize things like login and registration, update, delete, view one, view all, 
and validations. Its revolved around animes. Giving them a description and a rating from 1 - 100. Making the frontend nice and pretty will be my next thing
to look to update.